+++ startup/gui/editor.py
@@ -43,6 +43,7 @@ def __editorCreated( editor ) :
 	GafferSceneUI.SceneHistoryUI.connectToEditor( editor )
 	GafferSceneUI.EditScopeUI.connectToEditor( editor )
 	GafferUI._PlugVisibilityGadget.connectToEditor( editor )
+	GafferUI.UIEditor.connectToEditor ( editor )
 
 GafferUI.Editor.instanceCreatedSignal().connect( __editorCreated )
 GafferUI.CompoundEditor.nodeSetMenuSignal().connect( GafferUI.GraphBookmarksUI.appendNodeSetMenuDefinitions )