+++ Changes.md
@@ -6,6 +6,12 @@ Features
 
 - RandomPrimitiveVariable : Added new node for creating random per-face or per-vertex primitive variables with a variety of distributions.
 
+Improvements
+------------
+
+- GraphEditor : - Added <kbd>C</kbd> shortcut for setting node colors.
+- SceneView : Added keyboard shortcut <kbd>;</kbd> to toggle between the lookthrough (camera / light) view and the free (perspective / top / front / side) cameras. <kbd>Ctrl</kbd> + click on the toolbar camera icon will perform the same toggle.
+
 Fixes
 -----